pyproject.toml
[tool.poetry]
name = "invest-ohlcv-api-to-kafka"
version = "0.1.0"
description = ""
authors = ["datenraum <datenraum@me.com>"]

[tool.poetry.dependencies]
python = "^3.9"
rich = "^12.2.0"
click = "^8.1.2"
environs = "^9.5.0"
flake8 = "^4.0.1"
isort = "^5.10.1"
black = "^22.3.0"
orjson = "^3.6.7"
pandas = "^1.4.2"
PyYAML = "^6.0"
ratelimit = "^2.2.1"
icecream = "^2.1.2"
redis = "^4.2.2"
better-exceptions = "^0.3.3"
investpy = {git = "https://github.com/alvarobartt/investpy.git", rev = "master"}
kafka-python = "^2.0.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py36', 'py37', 'py38', 'py39', 'py310']
experimental_string_processing = true